pyproject.toml
[tool.poetry]
name = "njupt-always-in-dorm"
version = "0.1.0"
description = ""
authors = ["ArcticLampyrid <ArcticLampyrid@outlook.com>"]
readme = "README.md"
packages = [{include = "njupt_always_in_dorm", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26.4"
ddddocr = "^1.4.11"
requests = "^2.31.0"
pycryptodome = "^3.20.0"
schedule = "^1.2.1"
pytz = "^2024.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
njupt-always-in-dorm = "njupt_always_in_dorm.__main__:main"
